+++ nectar/CHANGELOG.md
@@ -23,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     If so, backup your data and then execute `nectar migrate-db run` to proceed with the migration.
 -   Add an optional `fund_amount` parameter to the `create-transaction` command.
     This allows users to redeem/refund `hbit` HTLCs that were funded with amounts different from what as agreed through the orderbook.
+    
+### Changed
+
+-   Replace Kraken mid market rate with a fixed rate of 1.
 
 ### Changed